+++ packages/language-support/src/tests/autocompletion/schemaBasedPatternCompletion.test.ts
@@ -1,6 +1,5 @@
 import { CompletionItemKind } from 'vscode-languageserver-types';
 import { testCompletions } from './completionAssertionHelpers';
-import { _internalFeatureFlags } from '../../featureFlags';
 
 const dbSchema = {
   labels: [
@@ -44,14 +43,6 @@ const dbSchema = {
 };
 
 describe('completeRelationshipType', () => {
-  beforeEach(() => {
-    _internalFeatureFlags.schemaBasedPatternCompletions = true;
-  });
-
-  afterEach(() => {
-    _internalFeatureFlags.schemaBasedPatternCompletions = false;
-  });
-
   test('Completion dont crash on missing db connection', () => {
     const query = 'MATCH (t:Trainer)-[r:';
 
@@ -480,7 +471,7 @@ RETURN [(p)-[:`;
     });
   });
 
-  test('Limitation: Does not handle union types, as they are not yet supported in symbol table ', () => {
+  test('Handles |-syntax', () => {
     const query = 'MATCH (x:Pokemon|Trainer)-[r:';
 
     testCompletions({
@@ -488,25 +479,128 @@ RETURN [(p)-[:`;
       dbSchema,
       computeSymbolsInfo: true,
       expected: [
-        /* Only these should be included
         { label: 'CHALLENGES', kind: CompletionItemKind.TypeParameter },
         { label: 'CATCHES', kind: CompletionItemKind.TypeParameter },
         { label: 'TRAINS', kind: CompletionItemKind.TypeParameter },
         { label: 'BATTLES', kind: CompletionItemKind.TypeParameter },
         { label: 'KNOWS', kind: CompletionItemKind.TypeParameter },
         { label: 'WEAK_TO', kind: CompletionItemKind.TypeParameter },
-         */
-        // For now, test bail working
+      ],
+      excluded: [
+        { label: 'UNRELATED_RELTYPE', kind: CompletionItemKind.TypeParameter },
+      ],
+    });
+  });
+
+  test('Handles &-syntax', () => {
+    const query = 'MATCH (x:Pokemon&Trainer)-[r:';
+
+    testCompletions({
+      query,
+      dbSchema,
+      computeSymbolsInfo: true,
+      expected: [
         { label: 'CHALLENGES', kind: CompletionItemKind.TypeParameter },
+      ],
+      excluded: [
         { label: 'CATCHES', kind: CompletionItemKind.TypeParameter },
         { label: 'TRAINS', kind: CompletionItemKind.TypeParameter },
         { label: 'BATTLES', kind: CompletionItemKind.TypeParameter },
         { label: 'KNOWS', kind: CompletionItemKind.TypeParameter },
         { label: 'WEAK_TO', kind: CompletionItemKind.TypeParameter },
         { label: 'UNRELATED_RELTYPE', kind: CompletionItemKind.TypeParameter },
+        { label: 'UNRELATED_RELTYPE', kind: CompletionItemKind.TypeParameter },
+      ],
+    });
+  });
+
+  test('Handles join &/|-syntax', () => {
+    //AND (&) gets evaluated first, so we get OR(AND(Pokemon, Trainer), Type)
+    const query = 'MATCH (x:Pokemon&Trainer|Type)-[r:';
+
+    testCompletions({
+      query,
+      dbSchema,
+      computeSymbolsInfo: true,
+      expected: [
+        { label: 'CHALLENGES', kind: CompletionItemKind.TypeParameter },
+        { label: 'STRONG_AGAINST', kind: CompletionItemKind.TypeParameter },
+        { label: 'WEAK_TO', kind: CompletionItemKind.TypeParameter },
       ],
       excluded: [
-        // { label: 'UNRELATED_RELTYPE', kind: CompletionItemKind.TypeParameter },
+        { label: 'CATCHES', kind: CompletionItemKind.TypeParameter },
+        { label: 'TRAINS', kind: CompletionItemKind.TypeParameter },
+        { label: 'BATTLES', kind: CompletionItemKind.TypeParameter },
+        { label: 'KNOWS', kind: CompletionItemKind.TypeParameter },
+        { label: 'UNRELATED_RELTYPE', kind: CompletionItemKind.TypeParameter },
+        { label: 'UNRELATED_RELTYPE', kind: CompletionItemKind.TypeParameter },
+      ],
+    });
+  });
+
+  test('Handles OR-syntax', () => {
+    const query = 'MATCH (x WHERE x:Pokemon OR x:Trainer)-[r:';
+
+    testCompletions({
+      query,
+      dbSchema,
+      computeSymbolsInfo: true,
+      expected: [
+        { label: 'CHALLENGES', kind: CompletionItemKind.TypeParameter },
+        { label: 'CATCHES', kind: CompletionItemKind.TypeParameter },
+        { label: 'TRAINS', kind: CompletionItemKind.TypeParameter },
+        { label: 'BATTLES', kind: CompletionItemKind.TypeParameter },
+        { label: 'KNOWS', kind: CompletionItemKind.TypeParameter },
+        { label: 'WEAK_TO', kind: CompletionItemKind.TypeParameter },
+      ],
+      excluded: [
+        { label: 'UNRELATED_RELTYPE', kind: CompletionItemKind.TypeParameter },
+      ],
+    });
+  });
+
+  test('Handles AND-syntax', () => {
+    const query = 'MATCH (x WHERE x:Pokemon AND x:Trainer)-[r:';
+
+    testCompletions({
+      query,
+      dbSchema,
+      computeSymbolsInfo: true,
+      expected: [
+        { label: 'CHALLENGES', kind: CompletionItemKind.TypeParameter },
+      ],
+      excluded: [
+        { label: 'CATCHES', kind: CompletionItemKind.TypeParameter },
+        { label: 'TRAINS', kind: CompletionItemKind.TypeParameter },
+        { label: 'BATTLES', kind: CompletionItemKind.TypeParameter },
+        { label: 'KNOWS', kind: CompletionItemKind.TypeParameter },
+        { label: 'WEAK_TO', kind: CompletionItemKind.TypeParameter },
+        { label: 'UNRELATED_RELTYPE', kind: CompletionItemKind.TypeParameter },
+        { label: 'UNRELATED_RELTYPE', kind: CompletionItemKind.TypeParameter },
+      ],
+    });
+  });
+
+  test('Handles joint AND/OR-syntax', () => {
+    //AND gets evaluated first, so we get OR(AND(Pokemon, Trainer), Type)
+    const query = 'MATCH (x WHERE x:Pokemon AND x:Trainer OR x:Type)-[r:';
+
+    testCompletions({
+      query,
+      dbSchema,
+      computeSymbolsInfo: true,
+      expected: [
+        { label: 'CHALLENGES', kind: CompletionItemKind.TypeParameter },
+        { label: 'STRONG_AGAINST', kind: CompletionItemKind.TypeParameter },
+        { label: 'WEAK_TO', kind: CompletionItemKind.TypeParameter },
+      ],
+      excluded: [
+        { label: 'CATCHES', kind: CompletionItemKind.TypeParameter },
+        { label: 'TRAINS', kind: CompletionItemKind.TypeParameter },
+        { label: 'BATTLES', kind: CompletionItemKind.TypeParameter },
+        { label: 'KNOWS', kind: CompletionItemKind.TypeParameter },
+        { label: 'UNRELATED_RELTYPE', kind: CompletionItemKind.TypeParameter },
+        { label: 'UNRELATED_RELTYPE', kind: CompletionItemKind.TypeParameter },
       ],
     });
   });