+++ packages/react-codemirror-playground/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @neo4j-cypher/react-codemirror-playground
 
+## 2.0.0-next.31
+
+### Patch Changes
+
+- Updated dependencies [5df1435]
+  - @neo4j-cypher/language-support@2.0.0-next.28
+  - @neo4j-cypher/react-codemirror@2.0.0-next.31
+
+## 2.0.0-next.30
+
+### Patch Changes
+
+- Updated dependencies [20b31b6]
+- Updated dependencies [68d92a8]
+  - @neo4j-cypher/language-support@2.0.0-next.27
+  - @neo4j-cypher/react-codemirror@2.0.0-next.30
+
 ## 2.0.0-next.29
 
 ### Patch Changes